Feat(Code Refactoring): refactored code for easy followup and readabillity
