--- actions/index.js
@@ -1,27 +1,0 @@
-import * as actionTypes from "./types";
-
-/* User Actions */
-export const setUser = user => {
-  return {
-    type: actionTypes.SET_USER,
-    payload: {
-      currentUser: user
-    }
-  };
-};
-
-export const clearUser = () => {
-  return {
-    type: actionTypes.CLEAR_USER
-  };
-};
-
-/* Channel Actions */
-export const setCurrentChannel = channel => {
-  return {
-    type: actionTypes.SET_CURRENT_CHANNEL,
-    payload: {
-      currentChannel: channel
-    }
-  }
-}
