--- actions/types.js
@@ -1,6 +1,0 @@
-/* User Action Types */
-export const SET_USER = "SET_USER";
-export const CLEAR_USER = "CLEAR_USER";
-
-/* Channel Action Types */
-export const SET_CURRENT_CHANNEL = "SET_CURRENT_CHANNEL";
