--- components/App.css
@@ -1,29 +1,0 @@
-html,
-body,
-.app {
-  height: 100vh;
-  background: #eee;
-  padding: 1em;
-}
-
-/* Messages.js */
-.messages {
-  height: 440px;
-  overflow-y: scroll;
-}
-
-/* MessageForm.js */
-.message__form {
-  position: fixed !important;
-  bottom: 1em;
-  margin-left: 320px !important;
-  left: 0;
-  right: 1em;
-  z-index: 200;
-}
-
-/* Message.js */
-.message__self {
-  border-left: 2px solid orange;
-  padding-left: 8px;
-}
