--- components/MetaPanel/MetaPanel.js
@@ -1,9 +1,0 @@
-import React, { Component } from "react";
-
-class MetaPanel extends Component {
-  render() {
-    return <div>MetaPanel</div>;
-  }
-}
-
-export default MetaPanel;
