+++ src/components/ColorPanel/ColorPanel.js
@@ -1,23 +1,22 @@
-import React, { Component } from 'react';
-import { Sidebar, Menu, Divider, Button, Icon } from "semantic-ui-react"
+import React from "react";
+import { Sidebar, Menu, Divider, Button } from "semantic-ui-react";
 
-
-class ColorPanel extends Component {
-   render() {
-      return (
-         <Sidebar
-            as={Menu}
-            icon="labeled"
-            inverted
-            vertical
-            visible
-            width="very thin"
-         >
-            <Divider />
-            <Button icon="left arrow icon" size="small" color="yellow" />
-         </Sidebar>
-      );
-   }
+class ColorPanel extends React.Component {
+  render() {
+    return (
+      <Sidebar
+        as={Menu}
+        icon="labeled"
+        inverted
+        vertical
+        visible
+        width="very thin"
+      >
+        <Divider />
+        <Button icon="angle double down" size="small" color="orange" />
+      </Sidebar>
+    );
+  }
 }
 
 export default ColorPanel;
