+++ src/components/Messages/ProgressBar.js
@@ -1,18 +1,16 @@
-import React, { Component } from 'react'
-import { Progress } from 'semantic-ui-react'
+import React from "react";
+import { Progress } from "semantic-ui-react";
 
-const ProgressBar = ({ uploadState, percentUploaded }) => (
-   uploadState && (
-      <Progress
-         className="progress__bar"
-         percent={percentUploaded}
-         progress
-         indicating
-         size="medium"
-         inverted
-      />
-   )
+const ProgressBar = ({ uploadState, percentUploaded }) =>
+  uploadState === "uploading" && (
+    <Progress
+      className="progress__bar"
+      percent={percentUploaded}
+      progress
+      indicating
+      size="medium"
+      inverted
+    />
+  );
 
-)
-
-export default ProgressBar
+export default ProgressBar;
